+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * bug: can't use coef_map with multiple statistics (thanks @sbw78)
 * bug: wrong number of stars w/ statistic='p.value' (thanks @torfason)
+* output='data.frame'. `extract` is no longer documented.
 
 # modelsummary 0.4.1